requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
pandas==2.1.3
numpy==1.24.3
tensorflow==2.15.0
streamlit==1.29.0
python-dotenv==1.0.0
google-generativeai==0.3.1